Add components for Privacy Policy Cookie Policy and Terms Of Use
