+++ client/src/modules/privacy-policy/components/CookiePolicy.jsx
@@ -1,5 +1,10 @@
 import React from 'react';
 
-const CookiePolicy = () => <div>CookiePolicy</div>;
+import Policy from './Policy';
+import CookiePolicyEn from './CookiePolicy.en.md';
+
+const CookiePolicy = () => (
+  <Policy dangerouslySetInnerHTML={{ __html: CookiePolicyEn }} />
+);
 
 export default CookiePolicy;
