--- client/src/modules/privacy-policy/components/PrivacyPolicy.scss
@@ -1,79 +1,0 @@
-@import 'partials';
-
-.privacy-policy {
-  width: 100%;
-  margin: 0 auto;
-  color: $text-normal;
-
-  &__intro-heading {
-    font-weight: 400;
-    text-align: center;
-    font-size: 52px;
-    padding-top: four-by(8);
-  }
-
-  &__intro-subheading {
-    font-weight: 400;
-    color: $text-normal;
-    text-align: center;
-    font-size: $s-font;
-    line-height: 1.5;
-    margin: four-by(4) 0;
-  }
-
-  &__info-wrapper {
-    padding: 0 four-by(4);
-    &:nth-child(even) {
-      background-color: $dark-orange;
-      color: $white;
-    }
-  }
-
-  &__info {
-    font-size: 16px;
-    padding: four-by(8) 0;
-    margin: 0 auto;
-    max-width: 70ch;
-  }
-
-  &__header {
-    font-weight: 500;
-    font-size: $xxxl-font;
-    padding: 0 0 four-by(2) 0;
-  }
-
-  &__content {
-    font-size: 16px;
-    line-height: 1.5;
-
-    a {
-      color: $white;
-      font-style: italic;
-      outline: 0;
-
-      &:hover {
-        color: $text-normal;
-      }
-    }
-  }
-
-  &__links {
-    max-width: 150px;
-    margin: 0 auto 100px;
-    text-align: center;
-    height: 40px;
-    outline: 0;
-
-    @include breakpoint($small) {
-      margin: 0 auto 70px;
-    }
-
-    @include breakpoint($large) {
-      margin: 0 auto 50px;
-    }
-
-    a {
-      text-decoration: none;
-    }
-  }
-}
